refactor: flatten stories
- removes / refactors some imports from deeper than src/
