+++ libs/spark/stories/colors.stories.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Meta } from '@storybook/react/types-6-0';
 import styled from 'styled-components';
-import { TextFormatter } from '../../src/utils/story-util';
+import { TextFormatter } from '../src/utils/story-util';
 import { Theme, useTheme } from '@material-ui/core';
 
 interface ColorBoxProps {
