+++ libs/spark/stories/icon-button.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Meta } from '@storybook/react/types-6-0';
-import { IconButton, IconButtonProps } from '../../src';
-import { Icons } from '../../src/elements/button';
+import { IconButton, IconButtonProps } from '../src';
+import { Icons } from './utils';
 
 export default {
   title: 'prenda-spark/IconButton',
