+++ libs/spark/stories/input.stories.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Story } from '@storybook/react';
 import { Meta } from '@storybook/react/types-6-0';
-
-import { Input, InputProps } from '../../src';
+import { Input, InputProps } from '../src';
 
 export default {
   title: 'prenda-spark/Input',
