+++ libs/spark/stories/menu.stories.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Meta } from '@storybook/react/types-6-0';
-
-import { Menu, MenuItemOptions } from '../../src';
+import { Menu, MenuItemOptions } from '../src';
 
 export default {
   title: 'prenda-spark/Menu',
