+++ libs/spark/stories/utils/Icons.tsx
@@ -1,0 +1,18 @@
+import * as React from 'react';
+import { createSvgIcon } from '@material-ui/core';
+
+// TODO: remove me when icons are bug-free
+// viewBox="0 0 24 24"
+const CaretDownIcon = createSvgIcon(
+  <path d="M6.69671 9.85355C6.50145 10.0488 6.50145 10.3654 6.69671 10.5607L11.6425 15.5064C11.7011 15.5651 11.7708 15.6061 11.8447 15.6295C12.0184 15.687 12.2174 15.6466 12.3556 15.5084L17.3033 10.5607C17.4986 10.3654 17.4986 10.0488 17.3033 9.85356L16.5962 9.14645C16.4009 8.95119 16.0844 8.95119 15.8891 9.14645L12 13.0355L8.11092 9.14645C7.91566 8.95118 7.59908 8.95118 7.40382 9.14645L6.69671 9.85355Z" />,
+  'SparkCaretDown'
+);
+const UserIcon = createSvgIcon(
+  <path d="M15.7128 13.7136C17.1058 12.6149 18 10.9119 18 9C18 5.68629 15.3137 3 12 3C8.6863 3 6.00001 5.68629 6.00001 9C6.00001 10.9119 6.89422 12.6149 8.28723 13.7136C5.36821 14.8898 3.22701 17.4152 3.01696 20.4999C2.9982 20.7754 3.22386 21 3.50001 21H4.50001C4.77615 21 4.99749 20.7754 5.02457 20.5006C5.31829 17.5204 8.21423 15 12 15C15.7858 15 18.6817 17.5204 18.9754 20.5006C19.0025 20.7754 19.2239 21 19.5 21H20.5C20.7761 21 21.0018 20.7754 20.9831 20.4999C20.773 17.4152 18.6318 14.8898 15.7128 13.7136ZM16 9C16 11.2091 14.2091 13 12 13C9.79087 13 8.00001 11.2091 8.00001 9C8.00001 6.79086 9.79087 5 12 5C14.2091 5 16 6.79086 16 9Z" />,
+  'SparkUser'
+);
+
+export const Icons = {
+  'caret-down': <CaretDownIcon />,
+  user: <UserIcon />,
+};
