+++ libs/spark/stories/utils/index.ts
@@ -1,0 +1,1 @@
+export { Icons } from './Icons';
